Added degreeday.py via upload
This is a python file I recently worked on to assist some customer requests at the Oklahoma Mesonet. I was able to take advantage of using dateframes and taking data from a website, which I felt like I learned a lot more in METR 3334. I took advantage of some preexisting code fragments and was able to create a working model to calculate degree day heat units with user inputs in Python for any Mesonet site for any time period.